Working minimum raggedness function.
Wrapping is done per paragraph, which is way faster than trying to
wrap the entire document at once.

Improve the cost function to consider the last line of a paragraph 0
cost. This allows orphans but has the benefit of keeping paragraphs
all around the desired width. Looks better for reading.